Fix DNS for most use cases
On iOS and macOS, domains like arsnova.click.local now work as
expected. However, on Linux and Android, the .local domain
is reserved for mDNS queries, which is a different protocol.
Hence, an investigation towards using mDNS on the "Server"
is necessary.

Simply opening `http://arsnova` in a browser works
consistently across OS.